beginnings of users support

+++ starter/views.py
@@ -1,18 +1,22 @@
 from django.shortcuts import render
-from django.http import Http404, JsonResponse, HttpResponse
+from django.http import Http404, Http403, JsonResponse, HttpResponse
 from .models import Missile, MissileDesigns
 from .forms import MissileForm
 from django.core import serializers
+from django.contrib.auth.decorators import login_required
 
 import json
 
 # Create your views here.
+@login_required
 def index(request):
     return render(request, 'starter/index.html')
 
+@login_required
 def designer(request):
     return render(request, 'starter/designer.html')
 
+@login_required
 def missile_designer(request):
     designs = MissileDesigns.objects.all()
     missiles  = [(design.design.pk, design.design.name) for design in designs]
@@ -22,8 +26,12 @@ def missile_designer(request):
     }
     return render(request, 'starter/missile-designer.html', context)
 
+
 def create_missile_design(request):
     if request.method == 'POST':
+        if not request.user.is_authenticated:
+            return Http403()
+
         #fields is "id_<field_name>" -> <field_data>
         #convert to "<field_name>" -> <field_data>
         convert_functions = {
@@ -56,6 +64,8 @@ def create_missile_design(request):
 
 def delete_missile_design(request):
     if request.method == 'POST':
+        if not request.user.is_authenticated:
+            return Http403()
         design_id = request.POST.get("design_id")
         print(design_id)
         missile = Missile.objects.get(pk=design_id).delete()
@@ -66,6 +76,8 @@ def delete_missile_design(request):
 
 def get_missile_design(request):
     if request.method == 'GET':
+        if not request.user.is_authenticated:
+            return Http403()
         design_id = request.GET.get("design_id")
         missile = Missile.objects.get(pk=design_id)
 
@@ -77,8 +89,10 @@ def get_missile_design(request):
     else:
          raise Http404()
 
+@login_required
 def gun_designer(request):
     raise Http404("Not implemented yet")
 
+@login_required
 def controller_designer(request):
     raise Http404("Not implemented yet")
